refactor: Pull input path processing out of compile method so compile can be called directly with no overhead if the input paths are known. Remove unused hooks feature
