+++ tests/approvals.spec.ts
@@ -2,6 +2,7 @@ import { globSync } from 'glob'
 import { rimraf } from 'rimraf'
 import { describe, expect, it } from 'vitest'
 import { compile } from '../src'
+import { processInputPaths } from '../src/input-paths/process-input-paths'
 import { isErrorOrCritical, LoggingContext, LogLevel } from '../src/logger'
 import { defaultPuyaOptions } from '../src/puya/options'
 import { normalisePath } from '../src/util'
@@ -19,10 +20,10 @@ describe('Approvals', async () => {
     ['O2', 'out/o2/[name]', { optimizationLevel: 2 }],
   ])('Compile %s', async (desc, outDir, puyaOptions) => {
     const logCtx = LoggingContext.create()
-    const result = await logCtx.run(() =>
-      compile({
-        paths: ['tests/approvals'],
-        outDir,
+    const result = await logCtx.run(() => {
+      const filePaths = processInputPaths({ paths: ['tests/approvals'], outDir })
+      return compile({
+        filePaths,
         dryRun: false,
         logLevel: LogLevel.Warning,
         skipVersionCheck: true,
@@ -35,8 +36,8 @@ describe('Approvals', async () => {
         outputArc56: true,
         outputSsaIr: true,
         ...puyaOptions,
-      }),
-    )
+      })
+    })
     it.each(contractFiles)('%s', (contractFilePath) => {
       const awst = result.awst?.filter((s) => s.sourceLocation.file === contractFilePath)
